move Cell Type to interface file
